oct-23 changes
