+++ src/main/java/org/milan/core/interfaces/Java8Interface.java
@@ -1,0 +1,19 @@
+package org.milan.core.interfaces;
+
+/**
+ * Use of Default and Static keyword to define non-abstract & utility method
+ *
+ * @author Milan Rathod
+ */
+public interface Java8Interface {
+
+    void testMethod();
+
+    default String testDefaultMethod() {
+        return "Test Default Method";
+    }
+
+    static String testStaticMethod() {
+        return "Test Static Method";
+    }
+}
