+++ src/main/java/org/milan/core/interfaces/TestInteface.java
@@ -1,0 +1,14 @@
+package org.milan.core.interfaces;
+
+/**
+ * Basic Java Interface Example
+ *
+ * @author Milan Rathod
+ */
+public interface TestInteface {
+
+    int i = 10;
+
+    void sayHello();
+
+}
