+++ src/main/java/org/milan/core/interfaces/TestInterfaceImpl.java
@@ -1,0 +1,16 @@
+package org.milan.core.interfaces;
+
+/**
+ * Implementation for {@link TestInteface}
+ */
+public class TestInterfaceImpl implements TestInteface {
+
+    public void sayHello() {
+        System.out.println("Hello");
+    }
+
+    public void sayBye() {
+        System.out.println("Bye");
+    }
+
+}
