--- src/main/java/org/milan/core/interfaces/abc.java
@@ -1,8 +1,0 @@
-package org.milan.core.interfaces;
-
-public interface abc {
-	int i = 10;
-
-	void sayHello();
-
-}
