--- src/main/java/org/milan/core/interfaces/p1.java
@@ -1,13 +1,0 @@
-package org.milan.core.interfaces;
-
-public class p1 implements abc {
-
-	public void sayHello() {
-		System.out.println("Hello");
-	}
-
-	public void sayBye() {
-		System.out.println("Bye");
-	}
-
-}
